app_data/requirements.txt
flask-wtf
flask-login
markupsafe
flask-migrate
PyMySQL
pylint
pytest-cov
sqlalchemy
